[ClangCL] Cleanup rest clang-cl warnings (#5847)
1. printf format mismatch.
2. avoid cast from CComPtr<IncludeHandlerVFSOverlayForTest> to
CComPtr<IDxcIncludeHandler>.
3. fix signed unsigned mismatch.
4. fix order of fields in constructor.
5. add override for override methods.
6. port
https://github.com/llvm/llvm-project/commit/01f4209631f32ccec686b232039238066a0866d9
7. add copy assignment operator to avoid
-Wdeprecated-copy-with-user-provided-copy
8. disable -Wignored-qualifiers for taef header WexTestClass.h.
9. remove unused fields.
10. add -clang-cl for hctbuild.

+++ tools/dxexp/dxexp.cpp
@@ -414,7 +414,7 @@ int main(int argc, const char *argv[]) {
     text_printf("Experimental shader model feature failed with unexpected "
                 "HRESULT 0x%08x.\n",
                 (unsigned int)hr);
-    json_printf("{ \"err\": \"0x%08x\" }", hr);
+    json_printf("{ \"err\": \"0x%08x\" }", (unsigned int)hr);
     json_printf("\n}\n");
     return 4;
   }
